% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.csc.matrix.R
\name{as.csc.matrix}
\alias{as.csc.matrix}
\title{converts to sparse matrix}
\usage{
as.csc.matrix(x, binary = FALSE, logical = FALSE, sort = FALSE)
}
\value{
dgCMatrix
}
\description{
converts to sparse matrix
}
