% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aslist_IBRAP_method.R
\name{as.list,methods-method}
\alias{as.list,methods-method}
\title{Method override for as.list function regarding methods S4 object}
\usage{
\S4method{as.list}{methods}(x)
}
\description{
Method override for as.list function regarding methods S4 object
}
