% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aslist.R
\name{as_list_methods}
\alias{as_list_methods}
\title{Method override for as.list function regarding methods S4 object}
\usage{
\S4method{as.list}{methods}(x)
}
\description{
Method override for as.list function regarding methods S4 object
}
