% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_matrix_transpose.R
\name{as_matrix_transpose}
\alias{as_matrix_transpose}
\title{converts to transposed matrix}
\usage{
as_matrix_transpose(mat)
}
\arguments{
\item{mat}{}
}
\value{
matrix
}
\description{
converts sparse matrix to normal transposed matrix for sparse data
}
