% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.intergation.R
\name{benchmark.integration}
\alias{benchmark.integration}
\title{Benchmarks the cluster assignmentws}
\usage{
benchmark.integration(
  object,
  batch,
  assays,
  reduction,
  result.names,
  n.components = 2
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{batch}{Character. Whcih column in the metadata dataframe contains the batches.}

\item{reduction}{Character. Which reduction(s) should be used to observe integration.}

\item{result.names}{Character. A vector of names for the benchmarking results.}

\item{n.components}{Numerical. How many components of the reduced embeddings should be used. Default = 2}

\item{assay}{Character. String containing indicating which assay to use}
}
\value{
Benchmarking scores for the supplied integrations
}
\description{
Adopting scones method of Average Silhouette Width determination between batches. A depleted value is superior.
}
\examples{
object <- benchmark.integration(object = object, 
                                          batch = 'original.project', 
                                          assays = c('SCT','SCRAN','SCANPY'), 
                                          reduction = c('pca_umap', 'pca_harmony_umap', 
                                                        'scanorama_umap', 'pca_bbknn_bbknn:umap',
                                                        'CCA_pca_umap'), 
                                          result.names = c('uncorrected', 'harmony', 
                                                           'scanorama', 'bbknn', 'cca'), 
                                          n.components = 2)

}
