% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_norm.R
\name{benchmark_norm}
\alias{benchmark_norm}
\title{Benchmarking of normalisation techniques}
\usage{
benchmark_norm(object, DR.assay, distance.method, n.k = 2:25)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{DR.assay}{Which dimensionality reduction assay should be utilised.}

\item{distance.method}{Which distance method should be used to calculate distance. Available: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'}

\item{n.k}{Either a single value, a list or a range of k to calculate.}
}
\description{
Benchmarking of normalisation techniques
}
\examples{
benchmark_norm(object = sce_object, DR.assay = 'pca', distance.method = 'euclidean', n.k=2:25)

}
