% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_clustering.R
\name{benchmarking_clustering}
\alias{benchmarking_clustering}
\title{Benchmark clustering results}
\usage{
benchmarking_clustering(
  object,
  components,
  reduction,
  dist.method = "euclidean",
  norm.methods,
  cluster.methods,
  ARI = FALSE,
  ground.truth
)
}
\arguments{
\item{components}{How many components to use from dimensionality reduction method}

\item{reduction}{Which reduction method to use.}

\item{dist.method}{Which distance method should be used to calculate distance. Available: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'}

\item{counts}{An expression MATRIX containing the raw counts of individual samples.}

\item{multi.norm}{TRUE/FALSE, whether multiple normalisation methods were used.}

\item{multi.cluster}{TRUE/FALSE, whether multiple clustering alogirthms were used.}
}
\description{
This function benchmarks the c;ustering methods.
}
\examples{
benchmarking_clustering(counts = counts.matrix, components = 1:19, reduction = 'pca', dist.method = 'euclidean', multi.norm = TRUE, multi.cluster = TRUE)

}
