% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_metadata.R
\name{cell_metadata}
\alias{cell_metadata}
\title{Generates cell-level metadata}
\usage{
cell_metadata(assay = counts, col.prefix = 'RAW')
}
\arguments{
\item{assay}{Counts matrix}

\item{col.prefix}{Which prefix to add to `'_total.counts' and '_total.feature'` as cell metdata column names}
}
\value{
Cell level metadata appended to metadata
}
\description{
Calculates total genes and total counts detected per cell
}
