% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_IBRAP_method.R
\name{colnames,IBRAP-method}
\alias{colnames,IBRAP-method}
\title{Method override for colnames function}
\usage{
\S4method{colnames}{IBRAP}(x, do.NULL = TRUE, prefix = "row")
}
\description{
Method override for colnames function
}
