% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_methods_method.R
\name{colnames_methods}
\alias{colnames_methods}
\title{Method override for colnames function}
\usage{
\S4method{colnames}{methods}(x, do.NULL = TRUE, prefix = "row")
}
\description{
Method override for colnames function
}
