% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sce_object.R
\name{create_sce_object}
\alias{create_sce_object}
\title{Produces SingleCellExperiment object}
\usage{
create_sce_object(counts, project.name)
}
\arguments{
\item{counts}{An expression MATRIX containing the raw counts of individual samples.}

\item{project.name}{Here name the individual sample/project.}
}
\description{
This function inputs an expression matrix into a singlecellexperiment object whilst calculating total RNA per cell and total feature appearance metadata.
}
\examples{
create_sce_object(counts = counts.matrix, project.name = "sample_1")

}
