% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIBRAPobject.R
\name{createIBRAPobject}
\alias{createIBRAPobject}
\title{Create an IBRAP class object}
\usage{
createIBRAPobject(counts = counts, original.project = 'project_1', method.name = 'RAW', meta.data = df, min.cells = 3, min.features = 200)
}
\arguments{
\item{counts}{Counts matrix}

\item{original.project}{Character string naming the project}

\item{method.name}{Character string naming the original method-assay. Default = 'RAW' `(Recommended)`}

\item{meta.data}{data.frame of extra metadata to append to generated dataframe. Warning: Must be in the same order and colnames}

\item{min.cells}{Numerical value of the minimum number of cells a gene should be present in}

\item{min.features}{Numerical value minimum features that should be present in a cell}
}
\value{
IBRAP S4 class object containing raw counts and metadata
}
\description{
Creates and produces project metadata into an IBRAP S4 class object
}
\examples{
object <- createIBRAPobject(counts = counts,
                                      meta.data = metadata_df,
                                      original.project = 'bmmc',
                                      method.name = 'RAW',
                                      min.cells = 3,
                                      min.features = 200)

}
