% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_dgCMatrix.R
\name{df_to_dgCMatrix}
\alias{df_to_dgCMatrix}
\title{converts to dgCMatrix}
\usage{
df_to_dgCMatrix(df, nsplit = 1000)
}
\arguments{
\item{mat}{}
}
\value{
mdgCMatrix
}
\description{
converts dataframe to  sparse matrix
}
