% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dollarsubset_IBRAP_method.R
\name{dollar_ibrap}
\alias{dollar_ibrap}
\title{Method override for `'$'` subset function regarding IBRAP S4 object}
\usage{
\S4method{$}{IBRAP}(x, name)
}
\description{
Method override for `'$'` subset function regarding IBRAP S4 object
}
