% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracketsubset_methods_method.R
\name{double_bracket_replace_methods}
\alias{double_bracket_replace_methods}
\title{Method override for `'[['` subset function regarding methods S4 object}
\usage{
\S4method{[[}{methods,ANY,ANY}(x, i, j, ...)
}
\description{
Method override for `'[['` subset function regarding methods S4 object
}
