% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracketsubset_IBRAP_method.R
\name{doublet_bracket_ibrap}
\alias{doublet_bracket_ibrap}
\title{Method override for subset method `'[['` for IBRAP object}
\usage{
\S4method{[[}{IBRAP,ANY,ANY}(x, i, j, ...)
}
\description{
Method override for subset method `'[['` for IBRAP object
}
