% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracketsubset_replace_IBRAP_method.R
\name{doublet_bracket_replace_ibrap}
\alias{doublet_bracket_replace_ibrap}
\title{Method override for colnames function}
\usage{
\S4method{[[}{IBRAP,ANY,ANY}(x, i, j) <- value
}
\description{
Method override for colnames function
}
