% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_metadata.R
\name{feature_metadata}
\alias{feature_metadata}
\title{Generates feature level metadata}
\usage{
feature_metadata(assay = counts, col.prefix = 'RAW')
}
\arguments{
\item{assay}{Counts matrix}

\item{col.prefix}{Which prefix to add to `'_total.counts' and '_total.cells'` as feature metdata column names}
}
\value{
Gene level metadata appended to metadata
}
\description{
Calculates total counts and cells detected per gene
}
