% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features_cells.R
\name{filter_features_cells}
\alias{filter_features_cells}
\title{Here filters the expression matrix}
\usage{
filter_features_cells(object, max.features, min.features, max.mt, min.cell)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{max.features}{Maximum n of features that should be present in a cell. This enables use to filter potential doublets}

\item{min.features}{Minimum n of features that should be present in a cell. This is important in omitting empty droplets}

\item{max.mt}{Maximum percetage of mitochondrial genes to be retained post-filtration. This helps to evade cells displaying mitochondrial damage during analysis.}

\item{min.cell}{The minimum number of cells a feature should be present in to be retained.}
}
\description{
Specific metadata variables can be filtered here. This importantly omits useless data , thus reducing computation time and increasing efficency.
For other metadata omission, please perform manually using the outlined example
}
\details{
Remember, you can view your available  metadata by specifying colData(sce_object)
}
