% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_IBRAP.R
\name{filter_IBRAP}
\alias{filter_IBRAP}
\title{Filters object according to cell metadata}
\usage{
filter_IBRAP(assay = counts, ... = RAW_total.features < max.features & RAW_total.counts > 200 & RAW_percent.mt < 8)
}
\arguments{
\item{...}{}

\item{assay}{IBRAP class object}
}
\value{
Filtered IBRAP object
}
\description{
Filters cells according to thresholds applied in the columns of object@metadata
}
