% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_other_features.R
\name{filter_other_features}
\alias{filter_other_features}
\title{Specific feature filtration}
\usage{
filter_other_features(object, features = list())
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{features}{Here specify the features you want to remove, i.e. c('MTOR', 'IGLC1')}
}
\description{
A defined feature list is omitted from the sce object
}
\examples{
filter_other_features(object = sce_object, features = c('MTOR', 'IGLC1'))
}
