% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_HVGs.R
\name{find_HVGs}
\alias{find_HVGs}
\title{Here performs Seurat::FindVariableFeatures}
\usage{
find_HVGs(object, nfeatures, ...)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{nfeatures}{How many features do you want to retain in as the most variable.}

\item{...}{Specify parameters to be passed to the FindVariableFeatures function.}
}
\description{
This method identifieds the most variable genes with a selection of three tests: vst. mvp, and disp
}
\examples{
find_HVGs(object = sce_object, nfeatures = 2000)

}
