% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_percentage_genes.R
\name{find_percentage_genes}
\alias{find_percentage_genes}
\title{Calculates the total percentage of genes matching a distinct character pattern.}
\usage{
find_percentage_genes(object, pattern = "^MT", column.name)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{pattern}{A regex pattern to identify genes}

\item{column.name}{Please name the column containing this feature information.}
}
\description{
This function calculates the percentage contribution of all transcripts associated with a specified regex pattern
}
