% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBRAP-object.R
\docType{class}
\name{IBRAP-object}
\alias{IBRAP-object}
\title{S4 class object that contains method-assays}
\description{
An S4 class object that contains assays and metadata.
}
\section{Slots}{

\describe{
\item{\code{methods}}{A list of methods class objects containing produced results}

\item{\code{sample_metadata}}{A data.frame containing cell level metadata}
}}

