% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.cluster.benchmarking.results.R
\name{integrate.cluster.benchmarking.results}
\alias{integrate.cluster.benchmarking.results}
\title{Integrates all cluster benchmarking into a single dataframe}
\usage{
integrate.cluster.benchmarking.results(object)
}
\arguments{
\item{object}{IBRAP S4 class object}
}
\value{
data frame of all clustering results separated by pipeline
}
\description{
Integrates all cluster benchmarking into a single dataframe
}
