% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isUnique.R
\name{isUnique}
\alias{isUnique}
\title{Are vector items unique}
\usage{
isUnique(vector = gene.names)
}
\arguments{
\item{vector}{}
}
\value{
Boolean result whether vector is unique
}
\description{
Identifies if a vector is unique
}
