% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_IBRAP_method.R
\name{merge_ibrap}
\alias{merge_ibrap}
\title{Method override for merge function regarding IBRAP S4 objects}
\usage{
\S4method{merge}{IBRAP,ANY}(x, y)
}
\description{
Method override for merge function regarding IBRAP S4 objects
}
