% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-object.R
\docType{class}
\name{methods-object}
\alias{methods-object}
\title{An S4 class object of method-assays}
\description{
An S4 class object that contains data matrices, HVGs, feature metadata, neighbourhood graphs, reduction embeddings, cluster assignments, and benchmarking results.
}
\section{Slots}{

\describe{
\item{\code{counts}}{Raw counts matrix in dgCMatrix format.}

\item{\code{normalised}}{Normalised matrix in dgCMatrix format.}

\item{\code{norm.scaled}}{HVG subset of normalised and scaled matrix in matrix format.}

\item{\code{highly.variable.genes}}{list of highly variable genes in character format.}

\item{\code{feature_metadata}}{data frame containing feature-level metadata.}

\item{\code{graphs}}{list of generated neighbourhood graphs (distances)}

\item{\code{computational_reductions}}{list of reductions to improve computational efficiency, i.e. PCA or dbMAP}

\item{\code{integration_reductions}}{list of integration output in reduction format}

\item{\code{visualisation_reductions}}{list of visualisation reductions, i.e. t-SNE, UMAP, etc.}

\item{\code{cluster_assignments}}{list of cluster assignment output. A dataframe is assigned to each method containing subsequent parameter changes.}

\item{\code{benchmarking_results}}{list of benchmarking results corresponding to the cluster assignment dataframes.}
}}

