% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_cpm.R
\name{normalisation_cpm}
\alias{normalisation_cpm}
\title{Here performs cpm package normalisation}
\usage{
normalisation_cpm(object, ...)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{...}{Here are functions to pass dowen to edgeR::cpm()}
}
\description{
Here performs cpm package normalisation
}
\examples{
normalisation_cpm(object = sce_object)
}
