% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_scran.R
\name{normalisation_scran}
\alias{normalisation_scran}
\title{Here performs scran package normalisation}
\usage{
normalisation_scran(
  object,
  max.cluster.size = 1000,
  scaling = NULL,
  log = TRUE,
  center_size_factors = TRUE
)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}
}
\description{
Scran normalisation comprises of pre-calculating clustering, computing the sum factors and then log normalising with the factors.
Scrans cutting edge is that it functions well wit high zero count data matrices.
}
\examples{
normalisation_scran(object = sce_object)
}
