% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_SCT.R
\name{normalisation_SCT}
\alias{normalisation_SCT}
\title{Here performs Seurat::SCTransfrom  normalisation}
\usage{
normalisation_SCT(object, ...)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{...}{Specify parameters to be passed to the SCTransform function.}
}
\description{
SCTransfrom, Seurats novel normalisation technique
}
\examples{
normalisation_SCT(object = sce_object)
}
