% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_SeuratLog.R
\name{normalisation_SeuratLog}
\alias{normalisation_SeuratLog}
\title{Here performs Seurat log transformation}
\usage{
normalisation_SeuratLog(object, ...)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{...}{Specify parameters to be passed to the NormalizeData function.}
}
\description{
This method logtransforms the data.
}
\examples{
normalisation_SeuratLog(object = sce_object)
}
