% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.bbknn.R
\name{perform.bbknn}
\alias{perform.bbknn}
\title{Performs BBKNN integration}
\usage{
perform.bbknn(
  object,
  assay,
  reduction,
  graph.name = "bbknn",
  batch,
  n_pcs = NULL,
  trim = NULL,
  n_trees = 10,
  use_faiss = TRUE,
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  save.object = TRUE
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{reduction}{Character. String defining the name of the reduction to provide for BBKNN. Default = NULL}

\item{graph.name}{Character. What should the BBKNN graph be named. Default = 'bbknn}

\item{batch}{Character. Column name in metadata indicating batch. Can be multiple.}

\item{n_pcs}{Numerical. Range of principal components to use. Default = NULL}

\item{trim}{Numerical. Trims the n of neighbours per cell to this value. Helps with population independence. Default = NULL}

\item{n_trees}{Numerical. Number of trees to generate in annoy forest. More trees provides higher precision at the cost of increased resource demand and run time. Default = 10}

\item{use_faiss}{Boolean. Uses faiss package to compute nearest neighbour, this improves run time at the cost of precision. Default = TRUE}

\item{set_op_mix_ratio}{Numerical. UMAP connectivity parameter between 0 and 1. controls the blen d between a connectivity matrix formed exclusively from mutual nearest neighbour pairs (0) and a union of all observed neighbour relationships with the mutual pairs emphasised (1). Default = 1.0}

\item{local_connectivity}{Numerical. How many nearest neighbours of each cell are assumed to be fully connected. Default = 1}
}
\value{
BBKNN connectivity graph contained in graphs in the indicated method-assays
}
\description{
Performs BBKNN integration on defined method-assays and reductions contained within. This is performed on reductions.
}
