% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.celltalker.R
\name{perform.celltalker}
\alias{perform.celltalker}
\title{Infers cell-to-cell communications.}
\usage{
add.cell.cycle(object = obj, assay = 'RAW', slot = 'counts')
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{A character string containing indicating which assay to use}

\item{slot}{String. indicating which slot within the assay should be sourced}

\item{clust.method}{String. can be either metadata or a dataframe stored in cluster_assignments}

\item{column}{String. name of the column to access in the clust.method}

\item{number_cells_required}{Number of cells per group required to perform
analysis of ligand/receptor interactions. Defaults to 100.}

\item{min_expression}{Minimum expression in counts to consider a ligand or
receptor for interactions analysis. A sensible default is set to 1000, but is
dataset dependent. This is meant to filter out lowly expressed ligands and
receptors.}

\item{max_expression}{Maxmium expression in counts to consider a ligand or
receptor for interactions analysis. A sensible default is set to 20000, but is
dataset dependent. This is meant to filter out ubiquitously expressed ligands
and receptors.}

\item{scramble_times}{Number of times to scamble ligand/receptor interactions to
create a background distribution for statistical comparison.}
}
\value{
dataframe of inferred cell-to-cell communications
}
\description{
Find ligand-receptor communcations between cell types.
}
