% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.dbmap.R
\name{perform.dbmap}
\alias{perform.dbmap}
\title{Performs dbMAP reduction}
\usage{
perform.dbmap(
  object,
  assay,
  slot = "normalised",
  n_components = NULL,
  n_neighbors = 15,
  reduction.save = "dbmap",
  save.object = TRUE
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{slot}{Character. String indicating which slot within the assay should be sourced}

\item{n_components}{Numerical. How many components should be produced, NULL enables automatic determination. Default = NULL}

\item{n_neighbors}{Numerical. How many neighbours are expected per cell. Default = 15}

\item{reduction.save}{Character. What should this reduction be saved as in computation_reduction. Default = 'dbmap'}

\item{save.object}{Boolean. Should the Anndata object use in python be saved under alt_objects. Default = TRUE}
}
\value{
dbMAP reductions contained within the computational_reduction list in the defined assays
}
\description{
Performs dbMAP reduction on defined method-assays. Unscaled but normalised data is automatically subset according to the identified HVGs for that method-assay
}
