% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.decontX.R
\name{perform.decontX}
\alias{perform.decontX}
\title{R package: celda, decontX function}
\usage{
perform.decontX(counts = counts)
}
\arguments{
\item{counts}{Counts matrix}

\item{z}{Cluster assignments for cells}

\item{maxIter}{Maximum number of iterations to be performed}

\item{delta}{Numeric Vector of length 2. Concentration parameters for the Dirichlet prior for the contamination in each cell. The first element is the prior for the native counts while the second element is the prior for the contamination counts. These essentially act as pseudocounts for the native and contamination in each cell. If estimateDelta = TRUE, this is only used to produce a random sample of proportions for an initial value of contamination in each cell. Then fit_dirichlet is used to update delta in each iteration. If estimateDelta = FALSE, then delta is fixed with these values for the entire inference procedure. Fixing delta and setting a high number in the second element will force decontX to be more aggressive and estimate higher levels of contamination at the expense of potentially removing native expression. Default c(10, 10).}

\item{estimateDelta}{Boolean. Whether to update delta at each iteration.}

\item{iterLogLik}{Integer. Calculate log likelihood every iterLogLik iteration. Default 10.}

\item{varGenes}{Integer. The number of variable genes to use in dimensionality reduction before clustering. Variability is calcualted using modelGeneVar function from the 'scran' package. Used only when z is not provided. Default 5000.}

\item{dbscanEps}{Numeric. The clustering resolution parameter used in 'dbscan' to estimate broad cell clusters. Used only when z is not provided. Default 1.}

\item{print.plot}{Logical. Should the UMAP plot displaying contaimination in each cell be printed? Default = FALSE}

\item{verbose}{Logical. Should function information be printed to the console? Default = FALSE}

\item{seed}{Integer. Passed to with_seed. For reproducibility. Default = 1234}
}
\value{
Doublet-omitted sparse matrix
}
\description{
Removes ambient RNA from datasets
}
\examples{
counts <- perform.decontx(counts = counts)

}
