% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.seurat.diffexp.all.R
\name{perform.diffexp.all}
\alias{perform.diffexp.all}
\title{Perform differential expression one cluster vs all}
\usage{
perform.diffexp.all(
  object,
  assay = NULL,
  test = "wilcox",
  identity = NULL,
  latent.vars = NULL,
  ...
)
}
\arguments{
\item{object}{An IBRAP class object}

\item{assay}{Character. Which assay within the IBRAP object to access. Default = NULL}

\item{test}{Character. Which test to use. Can be either: wilcox, bimod, roc, t, negbinom, poisson, LR, MAST, DESeq2. Please refer to Seurat::FindMarkers for more information.}

\item{identity}{Vector. A vector of cell identifiers to distinguish cells. Default = NULL}

\item{latent.vars}{Character. String(s) identifying which variables contained within the metadata to regress from the cells}

\item{...}{arguments to pass to Seurat::FindMarkers}
}
\value{
A dataframe containing differentially expression genes and other information
}
\description{
Performs differential expression on the assay differentiating the supplied identities, some methods enable variable regression, including: LR, negbinom, poisson or MAST. This function compares one cluster vs all others
}
\examples{

SCT_DE <- perform.seurat.diffexp.all(object = object, 
                                     assay = 'SCT', 
                                     test = 'MAST', 
                                     identity = object@sample_metadata$celltype, 
                                     latent.vars = 'original.project')
                                     
SCRAN_DE <- perform.seurat.diffexp.all(object = object, 
                                       assay = 'SCRAN', 
                                       test = 'MAST', 
                                       identity = object@sample_metadata$celltype, 
                                       latent.vars = 'original.project')
                                       
SCANPY_DE <- perform.seurat.diffexp.all(object = object, 
                                        assay = 'SCRAN', 
                                        test = 'MAST', 
                                        identity = object@sample_metadata$celltype, 
                                        latent.vars = 'original.project')

}
