% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.diffusion.map.R
\name{perform.diffusion.map}
\alias{perform.diffusion.map}
\title{Diffusion maps}
\usage{
perform.diffusion.map(
  object,
  assay,
  reduction,
  dims,
  n.dcs = 15,
  k = 15,
  diffmap.name.suffix = "",
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{reduction}{Character. String defining which reduction to supply to the clustering algorithm.}

\item{dims}{Numerical list. The number of dimensions to use for each reduction. This is supplied as a list respective to the order of reductions.}

\item{n.dcs}{Numerical. The number of diffusion components to produce. Default = 15}

\item{k}{Numerical. How many neighbours should be found per cell. A higher value tends to be more accurate. Default = 15}

\item{diffmap.name.suffix}{Character. What should be used as a suffix for diffmap}
}
\description{
Produces diffusion maps from previous reductions, i.e. PCA. Diffusion maps are known to better represent cellular trajectories in non-linear space
}
\examples{

                        

}
