% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.GO.enrichment.R
\name{perform.GO.enrichment}
\alias{perform.GO.enrichment}
\title{Plots slingshot results}
\usage{
perform.GO.enrichment(
  result,
  pval.cutoff = 0.001,
  gene.col = "gene",
  pval.col = "p_val",
  cluster.col = "cluster",
  nodeSize = 5,
  n.top.pathways = 10
)
}
\value{
A ggplot of the reduced cellular embbedings and trajectories.
}
\description{
Plots the results of the slingshot analysis using ggplots2
}
