% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.gsea.R
\name{perform.gsea}
\alias{perform.gsea}
\title{Returnas gene set enrichment scores for indivdiual cells}
\usage{
perform.gsea(object = obj)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{gene.sets}{Either a GeneSetCollection derived from escape::getGeneSets() or a list of genes with their grouping as the list item name. Default = escape::getGeneSets(library = 'H')}

\item{groups}{Numerical. The number of groups the cells should be split into. Default = 1000.}

\item{cores}{Numerical. The number of cores to commit to the calculations.}

\item{return_object}{Logical. Whether to return an updated IBRAP object}
}
\value{
IBRAP S4 class object containing module scores for each cell in the metadata
}
\description{
calculates gene set enrichment scores
}
