% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.harmony.R
\name{perform.harmony}
\alias{perform.harmony}
\title{Performs Harmony integration}
\usage{
perform.harmony(
  object,
  assay,
  batch,
  reduction = "pca",
  reduction.save.suffix = "",
  dims.use = NULL,
  print.harmony.plot = FALSE,
  seed = 1234,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{batch}{Character. A string of the column nmae that contains variables to regress.}

\item{reduction}{Character. String defining the name of the reduction to provide for HARMONY. Default = 'pca'}

\item{reduction.save.suffix}{Character. What should be appended to the end of harmony as the reduction name}

\item{dims.use}{Numerical. Number of dimensions of the provided reduction to input into harmony, NULL equates to all dimensions. Default = NULL}

\item{print.harmony.plot}{Boolean. Should the automatically generated plot be printed? Default = FALSE}

\item{seed}{Numeric. What should the seed be set as. Default = 1234}

\item{verbose}{Logical Should function messages be printed?}

\item{...}{Arguments to be passed to harmony::HarmonyMatrix}
}
\value{
PCA reductions contained within the computational_reduction list in the defined assays
}
\description{
Performs Harmony integration on defined method-assays and reductions contained within. This is performed on reductions.
}
\examples{

object <- perform.harmony(object = object, 
                          assay = c('SCRAN', 'SCT', 'SCANPY'), 
                          batch = 'original.project', 
                          reduction = c('pca'),  
                          max.iter.harmony = 100,
                          dims.use = list(NULL))

}
