% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.lvish.R
\name{perform.lvish}
\alias{perform.lvish}
\title{Performs LargeVis reduction}
\usage{
perform.lvish(
  object,
  assay,
  reduction = "pca",
  reduction.suffix = "",
  n.dim = NULL,
  n_components = 3,
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{reduction}{Character. String defining which reduction to supply to the LargeVis algorithm. Default = NULL}

\item{reduction.suffix}{Character. what should be appended to the end of lvish, cannot contain _.}

\item{n_components}{Numerical. How many tsne dimensions should be produced, if you are supplying graphs, only 2 dimensions can be produced. Default = 3}

\item{...}{Numerical. Arguments to be passed to uwot::lvish}

\item{n.dims}{Numerical. How many dimensions of the supplied reduction be used, Null equates to all dimensions. Default = NULL}
}
\value{
LargeVis reduction saved in the visualisation_reductions section in the supplied method-assays
}
\description{
Performs LargeVis reduction on defined method-assays and supplied reductions or graphs.
}
