% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.paga.trajectory.R
\name{perform.paga.trajectory}
\alias{perform.paga.trajectory}
\title{Create an IBRAP class object}
\usage{
createIBRAPobject(counts = counts, original.project = 'project_1', method.name = 'RAW', meta.data = df, min.cells = 3, min.features = 200)
}
\value{
IBRAP S4 class object containing raw counts and metadata
}
\description{
Creates and produces project metadata into an IBRAP S4 class object
}
