% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.pca.R
\name{perform.pca}
\alias{perform.pca}
\title{Performs PCA reduction}
\usage{
perform.pca(
  object,
  assay,
  slot = "norm.scaled",
  n.pcs = 50,
  reduction.save = "pca",
  save.plot = TRUE,
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{slot}{Character. String indicating which slot within the assay should be sourced}

\item{n.pcs}{Numerical. How many principal components should be produced. Default = 50}

\item{reduction.save}{Character. What should this reduction be saved as in computation_reduction. Default = 'pca'}

\item{...}{Arguments to be passed to PCAtools::pca}
}
\value{
PCA reductions contained within the computational_reduction list in the defined assays
}
\description{
Performs PCA reduction on defined method-assays. Data should be HVG subset, normalised and scaled (in the norm.scaled assay)
}
\examples{

object <- perform.pca(object = object, 
                      assay = c('SCT', 'SCRAN', 'SCANPY'), 
                      n.pcs = 50, 
                      reduction.save = 'pca')

}
