% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.reduction.kmeans.R
\name{perform.reduction.kmeans}
\alias{perform.reduction.kmeans}
\title{Performs Kmeans/PAM clustering on a reduction}
\usage{
perform.reduction.kmeans(
  object,
  assay,
  reduction = NULL,
  dims = NULL,
  k = NULL,
  cluster.df.name.suffix = "",
  method = "kmeans",
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{reduction}{Character. Which reduction(s) within the assay should be supplied. Default = NULL}

\item{dims}{Numerical. How many dimensions of the reduction should be supplied.Default = NULL}

\item{k}{Numerical. How many clusters should the algorithm identify, this can be a range. Default = NULL}

\item{method}{Which algorithm should be used, options: 'kmeans', 'pam'. Default = 'kmeans}

\item{...}{Arguments to be passed to either base::kmeans or cluster::pam, depending on the defined method}

\item{assignment.df.name}{Character. What to call the df contained in clusters.}
}
\value{
Cluster assignments using the list of resolutions provided contained within cluster_assignments under assignment.df.name
}
\description{
Performs Kmeans/PAM clustering on defined method-assays a defined reduction.
}
