% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.sc3.reduction.cluster.R
\name{perform.sc3.reduction.cluster}
\alias{perform.sc3.reduction.cluster}
\title{Performs SC3 clustering on reduced embeddings}
\usage{
perform.sc3.reduction.cluster(
  object,
  assay,
  reduction,
  dims,
  assignment.df.name,
  ks,
  n.core = 3
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{reduction}{Character. String defining which reduction to supply to the clustering algorithm.}

\item{dims}{Numerical. How many dimensions of the reduciton should be supplied, NULL equates to all.}

\item{assignment.df.name}{Character. What to call the df contained in clusters.}

\item{ks}{Numerical range. Number of clusters to identify, this can be a range, i.e. 5:10.}

\item{n.core}{Numerical. How many cores should be used to run SC3. Default = 3}
}
\value{
Cluster assignments using the list of resolutions provided contained within cluster_assignments under assignment.df.name
}
\description{
Performs SC3 clustering on defined method-assays and supplied reductions. SC3 is designed to function on count matrices, reduction may not always work well.
}
