% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.sc3.slot.cluster.R
\name{perform.sc3.slot.cluster}
\alias{perform.sc3.slot.cluster}
\title{Performs SC3 clustering on matrix slot}
\usage{
perform.sc3.slot.cluster(
  object,
  assay,
  slot,
  HVGs,
  cluster.df.name.suffix = "",
  ks,
  n.core = 3
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{slot}{Character. Which slot within the assay should be supplied.}

\item{HVGs}{Boolean. Should previously the count matrix be subset to highly variable genes.}

\item{ks}{Numerical range. Number of clusters to identify, this can be a range, i.e. 5:10.}

\item{n.core}{Numerical. How many cores should be used to run SC3. Default = 3}

\item{assignment.df.name}{Character. What to call the df contained in clusters.}
}
\value{
Cluster assignments using the list of resolutions provided contained within cluster_assignments under assignment.df.name
}
\description{
Performs SC3 clustering on defined method-assays slot.
}
