% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.scanorama.R
\name{perform.scanorama}
\alias{perform.scanorama}
\title{Performs Scanorama integration}
\usage{
perform.scanorama(
  object,
  assay,
  slot = "norm.scaled",
  split.by,
  n.dims = 50,
  reduction.save.suffix = "",
  batch_size = 5000,
  approx = TRUE,
  sigma = 15,
  alpha = 0.1,
  knn = 20,
  union = FALSE,
  seed = 12345
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{slot}{Character. String defining which slot in the assay to supply to Scanorama. Default = NULL}

\item{split.by}{Character. indicating the metadata column containing the batch to split the assay by.}

\item{n.dims}{Numerical. The number of Scanorama dimensions to be produced. Default = 50}

\item{reduction.save.suffix}{Character. Should a suffix be added to the end of scanorama, This cannot include underscores.}

\item{batch_size}{Numerical. The batch size used in the alignment vector computation. Useful when integrating large datasets. Default = 5000}

\item{approx}{Boolean. Use appoximate nearest neighbours within python, speeds up runtime. Default = TRUE}

\item{sigma}{Numerical. Correction smoothing parameter on Gaussian kernel. Default = 15}

\item{alpha}{Numerical. Alignment score minimum cutoff. Default = 0.1}

\item{knn}{Numerical. Number of nearest neighbors to use for matching. Default = 20}

\item{union}{Boolean. Should genes between datasets be intersected or not. Default = FALSE}

\item{seed}{Numerical. The seed to use when integrating these datasets. Default = 12345}
}
\value{
Scanorama reduction saved in the supplied method-assays
}
\description{
Performs Scanorama integration on defined method-assays and reductions contained within. This is performed on reductions.
}
