% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.scran.normalisation.R
\name{perform.scran}
\alias{perform.scran}
\title{Performs Scran normalisation, hvg selection, scaling and variance stabilisation and regression.}
\usage{
perform.scran(
  object,
  assay = "RAW",
  slot = "counts",
  batch = NULL,
  vars.to.regress = NULL,
  do.scale = FALSE,
  do.center = TRUE,
  new.assay.suffix = "",
  n.genes = 1500,
  max.cluster.size = 1000,
  center_size_factors = TRUE
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{slot}{Character. String indicating which slot within the assay should be sourced}

\item{batch}{Character. Which column in the metadata defines the batches. Default = NULL}

\item{vars.to.regress}{Character. Which column in the metadata should be regressed. Default = NULL}

\item{do.scale}{Boolean. Whether to scale the features variance. Default = TRUE}

\item{do.center}{Boolean. Whether to centre features to zero. Default = TRUE}

\item{new.assay.suffix}{Character. What should be added as a suffix for SCRAN}

\item{n.genes}{Numerical. Top number of genes to retain when finding HVGs. Default = 1500}

\item{max.cluster.size}{Numerical. When performing quickCluster, what is the maximum size the clusters can be. Default = 1000}

\item{center_size_factors}{Boolean Should size factor variance be centred. Default = TRUE}
}
\value{
Produces a new 'methods' assay containing normalised, scaled and HVGs.
}
\description{
A new method-assay is produced. Raw counts are normalised and HVGs identified using Scran
}
\examples{

object <- perform.scran(object = object, 
                        assay = 'RAW', 
                        slot = 'counts', 
                        vars.to.regress = 'RAW_total.counts', do.scale = T)

}
