% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.scrublet.R
\name{perform.scrublet}
\alias{perform.scrublet}
\title{Python module: scrublet}
\usage{
perform.scrublet(counts = counts, expected_doublet_rate = 0.025)
}
\arguments{
\item{counts}{Counts matrix}

\item{total_counts}{Total number of cells. NULL = automatically counts.}

\item{sim_doublet_ratio}{Number of doublets to simulate relative to observed}

\item{n_neighbors}{Expected number of neighbours per cell}

\item{expected_doublet_rate}{Expected percentage of doublets to be present in the dataset}

\item{stdev_doublet_rate}{Uncertainty in expected doublet rate}

\item{random_state}{Random state for doublet simulation, approximate nearest neighbour search, nd PCA/Truncated PCA}

\item{synthetic_doublet_umi_subsampling}{Sampling rate for UMIs in a cell when synthesising doublets}

\item{use_approx_neighbors}{Use approximate nearest neighbor method `(annoy)` for the KNN classifier}

\item{distance_metric}{Define distance metric for nearest neighbour calculation: 'angular', 'euclidean', 'manhattan', 'hamming', 'dot'.}

\item{get_doublet_neighbor_parents}{return the transcriptomes of the parent cells for simulated doublets}

\item{min_counts}{Minimum counts per cell}

\item{min_cells}{Minimum number of cells per gene}

\item{min_gene_variability_pctl}{Variability cutoff when deducing highly variable genes prior to PCA reduction}

\item{log_transform}{Log transforms the data}

\item{mean_center}{Should the dataset be centred around the mean}

\item{normalize_variance}{Should the genes have a total variance of 1}

\item{n_prin_comps}{Number of principal components to retain}

\item{svd_solver}{Which SVD solver to use: 'auto', 'full', 'arpack', 'randomized'.}
}
\value{
Doublet-omitted sparse matrix
}
\description{
Removes doublets from dataset.
}
