% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.sct.normalisation.R
\name{perform.sct.normalisation}
\alias{perform.sct.normalisation}
\title{Performs SCTransform}
\usage{
perform.sct.normalisation(
  object,
  assay,
  slot,
  new.assay.name = "SCT",
  do.scale = TRUE,
  do.center = TRUE,
  n.genes = 1500,
  min_cells = 3,
  save.seuratobject = TRUE,
  vars.to.regress = NULL,
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{A character string containing indicating which assay to use}

\item{slot}{String indicating which slot within the assay should be sourced}

\item{new.assay.name}{Character. Name of new method-assay to be produced. Default = 'SCT'}

\item{do.scale}{Whether to scale residuals to have unit variance; default is FALSE}

\item{do.center}{Whether to center residuals to have mean zero; default is TRUE}

\item{n.genes}{Numerical value of how many highly variable genes should be retained. Default = 1500}

\item{min_cells}{Numerical value of minimum cells required for a gene to not be filtered. Default = 3}
}
\value{
Produces a new 'methods' assay containing normalised, scaled and HVGs.
}
\description{
A new method-assay is produced. Raw counts are normalised and HVGs identified using SCTransform
}
