% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.sctype.R
\name{perform.sctype}
\alias{perform.sctype}
\title{Performs sctype}
\usage{
perform.sctype(
  object,
  assay = "RAW",
  slot = "counts",
  db =
    "https://raw.githubusercontent.com/IanevskiAleksandr/sc-type/master/ScTypeDB_full.xlsx",
  tissue
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{slot}{Character. String defining which slot in the assay to supply to Scanorama. Default = NULL}

\item{db}{Character. String indicating where to collect database from. So far we only use the standard scType reference.}

\item{tissue}{Character.String which tissue to subset.}
}
\value{
cell type annotation in relation to the clusterign categories that were provided
}
\description{
Performs sctype
}
