% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.seurat.diffexp.R
\name{perform.seurat.diffexp}
\alias{perform.seurat.diffexp}
\title{Plots slingshot results}
\usage{
perform.seurat.diffexp(object, assay, test, identity, latent.vars = NULL, ...)
}
\value{
A ggplot of the reduced cellular embbedings and trajectories.
}
\description{
Plots the results of the slingshot analysis using ggplots2
}
