% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.slingshot.trajectory.R
\name{perform.slingshot.trajectory}
\alias{perform.slingshot.trajectory}
\title{Performs Slingshot trajectory inference}
\usage{
perform.slingshot.trajectory(
  object,
  reduction,
  assay,
  clust.method,
  column,
  start.clus = NULL,
  end.clus = NULL,
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{reduction}{Character. String defining which reduction to supply to the clustering algorithm. Default = NULL}

\item{assay}{Character. String containing indicating which assay to use}

\item{clust.method}{Character. Which cluster method should be used utilised from clustering results, if `'metadata'` is supplied, you will access the metadata.}

\item{column}{Character. Which column within the isolated clust.method should be used to define cell type annoation.}

\item{start.clus}{Character. Which cluster should start the trajectory, if NULL then slingshot will attempt to discover this. Default = NULL}

\item{end.clus}{Character. Which cluster should end the trajectory, if NULL then slingshot will attempt to discover this. Default = NULL}

\item{...}{arguments to be passed to slingshot::slingshot}
}
\value{
A SingleshotDataSet class results object containing cellular lineages/curves
}
\description{
Generates slingshot trajectory inference on the defined reduction and clustering
}
\examples{

traj_res <- perform.slingshot.trajectory(object = object, reduction = 'pca_umap', 
                                         assay = 'SCT', clust.method = 'metadata', 
                                         column = 'celltype')
                                         
plot.slingshot(result = traj_red, object = traj, assay = 'SCT', relevant = F, Pseudotime = T)

}
