% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.tpm.normalisation.R
\name{perform.tpm.normalisation}
\alias{perform.tpm.normalisation}
\title{Performs TPM normalisation}
\usage{
perform.tpm.normalisation(
  object,
  assay = "RAW",
  slot = "counts",
  n.genes = 1500,
  do.scale = FALSE,
  do.center = TRUE,
  vars.to.regress = NULL,
  new.assay.name = "TPM"
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{slot}{Character. String indicating which slot within the assay should be sourced}

\item{n.genes}{Numerical. Top number of genes to retain when finding HVGs. Default = 1500}

\item{do.scale}{Boolean. Whether to scale the features variance. Default = TRUE}

\item{vars.to.regress}{Character. Which column in the metadata should be regressed. Default = NULL}

\item{new.assay.name}{Character. What should the new assay be called. Default = 'SCRAN'}

\item{do.centre}{Boolean. Whether to centre features to zero. Default = TRUE}
}
\value{
Produces a new 'methods' assay containing normalised, scaled and HVGs.
}
\description{
A new method-assay is produced. Raw counts are normalised and HVGs identified using TPM and Scran, respectively.
}
