% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.tsne.R
\name{perform.tsne}
\alias{perform.tsne}
\title{Performs t-SNE reduction}
\usage{
perform.tsne(
  object,
  assay,
  reduction = NULL,
  reduction.name.suffix = "",
  n.dims = NULL,
  n_components = 3,
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{reduction}{Character. String defining which reduction to supply to the t-SNE algorithm. Default = NULL}

\item{reduction.name.suffix}{Character. What should be appended to the end of tsne as the reduction name.}

\item{n.dims}{Numerical. How many dimensions of the supplied reduction be used, Null equates to all dimensions. Default = NULL}

\item{n_components}{Numerical. How many tsne dimensions should be produced, if you are supplying graphs, only 2 dimensions can be produced. Default = 3}

\item{...}{Numerical. Arguments to be passed to ProjectionBasedClustering::tSNE}
}
\value{
UMAP reduction saved in the visualisation_reductions section in the supplied method-assays
}
\description{
Performs t-SNE reduction on defined method-assays and supplied reductions or graphs.
}
