% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performPCA.R
\name{performPCA}
\alias{performPCA}
\title{Principal Component Analysis (PCA)}
\usage{
performPCA(object, ...)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{...}{Specify parameters to be passed to the pca function from pcaMethods.}
}
\description{
Principal Component Analysis (PCA)
}
\examples{
performPCA(object = sce_object, unwanted.variance = 'percent.mt', scale = TRUE, centre = TRUE)
}
