% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dot.plot.R
\name{plot.dot.plot}
\alias{plot.dot.plot}
\title{Plots a dot plot of gene expression}
\usage{
\method{plot}{dot.plot}(object, assay, slot = "normalised", clust.method, column, features)
}
\arguments{
\item{object}{An object of IBRAP class}

\item{assay}{Character. Which assay to access}

\item{slot}{Character. Which expression matrix slot should be used. Default = 'normalised'}

\item{clust.method}{Character. Which cluster method should be used, metadata accesses objects metadata}

\item{column}{Character. Which column to access in the defined clust.method}

\item{features}{Character. Which features to plot}
}
\description{
Creates a dot plot of gene expression
}
