% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.features.R
\name{plot.features}
\alias{plot.features}
\title{Plot of reduced dimensions and features}
\usage{
\method{plot}{features}(
  object,
  assay,
  slot,
  reduction,
  features,
  order = TRUE,
  percentile = c(0, 1),
  pt_size = 0.5,
  cells = NULL,
  colours = c("#C0C0C0", "#4169E1")
)
}
\arguments{
\item{object}{An IBRAP S4 class object}

\item{assay}{Character. Which assay within the object to access}

\item{slot}{Character. Which expression matrix would you like to access. Default = 'normalised'}

\item{reduction}{Character. Which reduction should be used for the plot}

\item{features}{Character Which features should be plotted}

\item{order}{Boolean. Should datapoints be plotted in order of expression intensity. Default = TRUE}

\item{pt_size}{Numeric. what size should the inidividual plot sizes be}

\item{cells}{Numeric. Which cells should be subset for plotting, Default = NULL}

\item{colours}{Vector. The first value will represent the lower values and the second the highest}

\item{percentile.}{Numerical. What percentile of datapoint expression should be plotted. Default = c(0.1,0.9)}
}
\value{
A plot of reduced dimensions annotated with assignments
}
\description{
Plot of reduced dimensions and features
}
