% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.QC.scatter.R
\name{plot.QC.scatter}
\alias{plot.QC.scatter}
\title{Plots two QC metrices in scatter format}
\usage{
plot.QC.vln(object = obj, metadata.columns = `c('RAW_total.features', 'RAW_total.counts')`, split.by = 'original.project')
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{x}{Character string of which metadata column to subset for the x axis}

\item{y}{Character string of which metadata column to subset for the y axis}

\item{split.by}{Character string of which metadata column to separate cells by}
}
\value{
A Violin plot
}
\description{
Plots two QC metric in scatter format
}
