% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.QC.vln.R
\name{plot.QC.vln}
\alias{plot.QC.vln}
\title{Plots a given QC metric}
\usage{
plot.QC.vln(object = obj, metadata.columns = `c('RAW_total.features', 'RAW_total.counts')`, split.by = 'original.project')
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{metadata.columns}{A character string or vector of character strings of which metadata columns to subset and plot}

\item{split.by}{A character string indicating which metadata column should be used to divide samples}
}
\value{
A Violin plot
}
\description{
Plots a defined QC metric in violin format
}
