% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reduced.dim.interactive.R
\name{plot.reduced.dim.interactive}
\alias{plot.reduced.dim.interactive}
\title{Interactive plot of reduced dimensions and labels}
\usage{
\method{plot}{reduced.dim.interactive}(
  object,
  reduction,
  assay,
  clust.method,
  column,
  pt.size = 5,
  dimensions,
  cells = NULL,
  ...
)
}
\arguments{
\item{object}{An IBRAP S4 class object}

\item{reduction}{Character. Which reduction to use to display points}

\item{assay}{Character. Which assay within the object to access}

\item{clust.method}{Character. Which clustering method to access, supply the name of any dataframe contained in the cluster_assignments sections. If you wish to access metadata, just specify 'metadata'}

\item{column}{Character. Which column to access within the supplied clust.column}

\item{pt.size}{Numeric. What should the point size be}

\item{dimensions}{Numeric. How many dimensions should be displayed, can be either 2 or 3}

\item{cells}{Numeric. Which cells should be subset for plotting, Default = NULL}
}
\value{
An interactive plot of reduced dimensions annotated with assignments
}
\description{
Interactive plot of reduced dimensions and labels
}
