% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reduced.dim.R
\name{plot.reduced.dim}
\alias{plot.reduced.dim}
\alias{plot.reduced.dim.}
\title{Plot of reduced dimensions and labels}
\usage{
\method{plot}{reduced.dim}(object, reduction, assay, clust.method, column, pt.size = 5, cells = NULL)
}
\arguments{
\item{object}{An IBRAP S4 class object}

\item{reduction}{Character. Which reduction to use to display points}

\item{assay}{Character. Which assay within the object to access}

\item{clust.method}{Character. Which clustering method to access, supply the name of any dataframe contained in the cluster_assignments sections. If you wish to access metadata, just specify 'metadata'}

\item{column}{Character. Which column to access within the supplied clust.column}

\item{pt.size}{Numeric. What should the point size be}

\item{cells}{Numeric. Which cells should be subset for plotting, Default = NULL}
}
\value{
A plot of reduced dimensions annotated with assignments
}
\description{
Plot of reduced dimensions and labels
}
