% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.slingshot.R
\name{plot.slingshot}
\alias{plot.slingshot}
\title{Plots slingshot results}
\usage{
\method{plot}{slingshot}(
  result,
  clusters = NULL,
  lineages = FALSE,
  pt_size = 0.1,
  line_size = 0.1,
  relevant = TRUE,
  Pseudotime = FALSE,
  Expression = FALSE,
  object = NULL,
  assay = NULL,
  slot = "norm.scaled",
  feature = NULL,
  ...
)
}
\arguments{
\item{result}{A SlingshotDataSet class results object}

\item{clusters}{Vector. A vector of clusters to use in ggplot, if NULL the clusters used to generate results will be used. Default = NULL}

\item{lineages}{Boolean. Should the direct lineages be applied or curving, TRUE = curves, FALSE = lineages. Default = FALSE}

\item{pt_size}{Numerical. What size should the cell points be. Default = 0.1}

\item{line_size}{Numerical. What size should the lineage lines be. Default = 0.1}

\item{relevant}{Should only pseudotime relevant cells be plotted. Default = TRUE}

\item{...}{arguments to be passed to ggplot::geom_point}

\item{title}{Character. Name the graph. Default = NULL}

\item{lab.clusters}{Boolean. Whether the clusters within a lineage should be labelled, note this only works for lineages and not curves. Default = TRUE}
}
\value{
A ggplot of the reduced cellular embbedings and trajectories.
}
\description{
Plots the results of the slingshot analysis using ggplots2
}
