% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.variance.R
\name{plot.variance}
\alias{plot.variance}
\title{Plot reduction explained variance}
\usage{
\method{plot}{variance}(object, assay, reduction, ...)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{reduction}{Character. Which reductions to access within the supplied assays}

\item{...}{Arguments to be passed to ggpubr::ggarrange}
}
\value{
Plots showing the supplied reductions explained variance
}
\description{
Shows the explained variance for a reduction for selecting how many dimensions to use downstream
}
