% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vln.R
\name{plot.vln}
\alias{plot.vln}
\title{Plot of violin plot of defined features}
\usage{
\method{plot}{vln}(
  object,
  assay,
  slot,
  features,
  group.by,
  title = NULL,
  xlab = "group",
  ylab = "expression"
)
}
\arguments{
\item{object}{An IBRAP S4 class object}

\item{assay}{Character. Which assay within the object to access}

\item{slot}{Character. Which expression matrix would you like to access. Default = 'normalised'}

\item{features}{Character. Which features should be plotted, it is not recommended to exceed 3 but can be more if your screen size is larger.}

\item{group.by}{Vector. What should the cell types be split by.}

\item{title}{Character. What should be the overall title of the plot. Default = NULL}

\item{xlab}{Character. What should by the x axis title. Default = 'group'}

\item{ylab}{Character. What shoul be the y axis title. Default = 'expression'}
}
\value{
A violin plot of defined gene expression and groups
}
\description{
Plot of violin plot of defined features
}
