% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benchmarking.R
\name{plot_benchmarking}
\alias{plot_benchmarking}
\title{Benchmark metric visualisation tool for clustering}
\usage{
plot_benchmarking(object, norm.method, clust.method, ARI)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{norm.method}{Identify which normalisation method you would like to access.}

\item{clust.method}{Which cliuster method would you like to assess?}
}
\description{
Here will display average silhouette width, dunn index, cluster connectivity, and adjusted rand index.
}
