% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_dr.R
\name{plot_cluster_dr}
\alias{plot_cluster_dr}
\title{Dimensionality reduction plot.}
\usage{
plot_cluster_dr(
  object,
  reduction = "umap",
  pt.size = 0.5,
  norm.method,
  cluster.method,
  cluster.column,
  interactive = FALSE
)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{reduction}{Which dimensionality reduction assay should be utilised?}

\item{pt.size}{What size should be the data points be.}

\item{norm.method}{Which normalisation method wouyld you like to access?}

\item{cluster.method}{Which clustering method would you like to access?}

\item{cluster.column}{Which column within these results would you like to group by?}

\item{multi.norm}{TRUE/FALSE were multiple normalisation techniques performed?}

\item{multi.cluster}{TRUE/FALSE were multiple clustering techniques performed?}
}
\description{
Here two dimensions will be displayed.
}
\examples{
plot_cluster_dr(object = sce, reduction='umap', pt.size=0.5, multi.norm=TRUE, multi.cluster=TRUE, norm.method='scran', cluster.method='sc3', cluster.column='SCT_snn_res.0.4')

}
