% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reddim_var.R
\name{plot_reddim_var}
\alias{plot_reddim_var}
\title{Benchmark clustering results}
\usage{
plot_reddim_var(object, norm.method, reduction, n.dim, cex.names = 0.6)
}
\arguments{
\item{norm.method}{list of normalisation methods used.}

\item{reduction}{Which reduction method to use.}

\item{n.dim}{How many dimensions should be used.}

\item{cex.names}{this value indicates label size}

\item{counts}{An expression MATRIX containing the raw counts of individual samples.}
}
\description{
This function benchmarks the c;ustering methods.
}
