% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.for.dittoSeq.R
\name{prepare.for.dittoSeq}
\alias{prepare.for.dittoSeq}
\title{Create a SCE object to use for creating plots in dittoSeq}
\usage{
prepare.for.dittoSeq(
  object,
  assay,
  slot = "normalised",
  clust.method,
  column,
  reduction = NULL,
  ...
)
}
\arguments{
\item{object}{An IBRAP S4 class object}

\item{assay}{Character. Which assay within the object to access}

\item{reduction}{Character. Which reduction to use to display points}

\item{clust.method.}{Character. Whcih cluster assignments should be added to the metadata. the dataframe contained in `object@sample_metadata` is alread contained in the object. NULL will mean only sample_metadata will be in the returned object}

\item{pt.size}{Numeric. What should the point size be}

\item{cells}{Numeric. Which cells should be subset for plotting, Default = NULL}
}
\value{
A SCE object containing
}
\description{
Create a SCE object to use for creating plots in dittoSeq
}
