% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.reticulate.R
\name{prepare.reticulate}
\alias{prepare.reticulate}
\title{Installs or identifies if python modules are installed}
\usage{
prepare.reticulate()
}
\description{
This function checks if IBRAPs dependent
}
