% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_tSNE.R
\name{produce_tSNE}
\alias{produce_tSNE}
\title{Uniform Manifold pproximation and Projection (UMAP)}
\usage{
produce_tSNE(object, norm.method, init.dims, reduction = "pca", ...)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{norm.method}{Please identify the normalisation techniques.}

\item{init.dims}{How many dimensions should be used?}

\item{reduction}{Which reduction method would you like to use?}

\item{...}{Further parameters to pass down to the umap function from tSNE}
}
\description{
performs non-linear dimensionality reduction using UMAP
}
\examples{
produce_umap(object = sce_object, multi.norm = TRUE, multi.cluster = TRUE)

}
