% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_umap.R
\name{produce_umap}
\alias{produce_umap}
\title{Uniform Manifold pproximation and Projection (UMAP)}
\usage{
produce_umap(object, norm.methods, reduction = "pca", n.dim, ...)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{...}{Further parameters to pass down to the umap function from uwot.}

\item{multi.norm}{TRUE/FALSE, were multiple normalisation techniques used?}

\item{multi.cluster}{TRUE/FALSE, were multiple clustering techniques used?}
}
\description{
performs non-linear dimensionality reduction using UMAP
}
\examples{
produce_umap(object = sce_object, multi.norm = TRUE, multi.cluster = TRUE)

}
