% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read10X_output.R
\name{Read10X_output}
\alias{Read10X_output}
\title{Produce counts matrix from CellRanger output}
\usage{
Read10X_output(directory, matrix.file, genes.file, barcodes.file, make.feat.unique)
}
\arguments{
\item{directory}{Absolute pathway to directory containing the three files}

\item{matrix.file}{Specify .mtx file name. Default = matrix.mtx}

\item{genes.file}{Specify file name for genes. Default = genes.tsv}

\item{barcodes.file}{Specify file name for barcodes. Default = barcodes.tsv}

\item{make.feat.unique}{Should non-unique gene names be made unique.}
}
\value{
Sparse matrix of counts
}
\description{
Concatenates the three output files of CellRanger alignment.
}
\examples{
Read10X_output(directory = 'path/to/folder/', matrix.file = 'matrix.mtx', genes.file = 'genes.tsv', barcodes.file = 'barcodes.tsv', make.feat.unique = TRUE)

}
