% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.hvgs.R
\name{remove.hvgs}
\alias{remove.hvgs}
\title{Remove highly variable genes}
\usage{
remove.hvgs(object, assay, hvgs.omit)
}
\arguments{
\item{object}{An IBRAP S4 class object}

\item{assay}{Character vector. Which assays to access}

\item{hvgs.omit}{List of character vectors. A list of genes to remove from HVG list.}
}
\description{
This function removes highly variable genes from method-assays
}
