% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceResults.R
\name{replaceResults}
\alias{replaceResults}
\title{Shows the contents in your IBRAP object}
\usage{
replaceResults(object, assay, item.to.replace, replacement.item)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String showing which assay to access}

\item{item.to.replace}{Character. The name of the item to replace in the object.}

\item{replacement.item}{Supply the item to replace the item with.}
}
\value{
Prints out the contents of the supplied assays
}
\description{
Shows the contents in your IBRAP object
}
