% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames_IBRAP_method.R
\name{rownames,IBRAP-method}
\alias{rownames,IBRAP-method}
\title{Method override for rownames function}
\usage{
\S4method{rownames}{IBRAP}(x, do.NULL = TRUE, prefix = "row")
}
\description{
Method override for rownames function
}
