% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames_methods_method.R
\name{rownames,methods-method}
\alias{rownames,methods-method}
\title{Method override for rownames function}
\usage{
\S4method{rownames}{methods}(x, do.NULL = TRUE, prefix = "row")
}
\description{
Method override for rownames function
}
