% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.IBRAP.rshiny.R
\name{run.IBRAP.rshiny}
\alias{run.IBRAP.rshiny}
\title{Rshiny application initiator}
\usage{
run.IBRAP.rshiny(...)
}
\arguments{
\item{...}{arguments to pass to shiny::runApp}
}
\description{
Initiates the Rshiny application for data interpretation
}
