% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_IBRAP_clustering_browser.R
\name{Run_IBRAP_clustering_browser}
\alias{Run_IBRAP_clustering_browser}
\title{Benchmarking of normalisation techniques}
\usage{
Run_IBRAP_clustering_browser()
}
\description{
This function activates the shiny app for investigating the c;ustering results
}
\examples{
benchmark_norm(object = sce_object, DR.assay = 'pca', distance.method = 'euclidean', n.k=2:25)

}
