% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC3_cluster.R
\name{SC3_cluster}
\alias{SC3_cluster}
\title{SC3 consensus clustering}
\usage{
SC3_cluster(object, norm.methods, ks, n.core = 2)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{ks}{Specify k value (number of total clusters)}

\item{n.core}{SC3 is computationally demanding, and thus may require multiple cores. please specify here. For a local computer use (total cores - 1), in a hpc use all cores}

\item{multi.method}{TRUE/FALSE, were multiple normnalisation techniques used?}
}
\description{
Functions by using hierachial clustering to determine a consensus across kmeans calculations
}
\examples{
SC3_cluster(object = sce_object, multi.method = TRUE, ks = 12, n.core = 2)

}
