% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Cluster.R
\name{sc_Cluster}
\alias{sc_Cluster}
\title{An integrated clustering step}
\usage{
sc_Cluster(
  object,
  norm.method,
  cluster.method,
  res = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 1.1, 1.2, 1.3, 1.4, 1.5),
  prune.SNN = 0,
  nn.method = "annoy",
  annoy.metric = "euclidean",
  nn.eps = 0,
  num.core,
  ks,
  reduction,
  dims.use,
  nl.reduction.method,
  benchmark,
  dist.method = "euclidean",
  ARI,
  ground.truth
)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{norm.method}{Specify which normalisation method you would like to process}

\item{cluster.method}{Identify the cluster methods used}

\item{num.core}{How many cores should be used during SC3 processesing.}

\item{reduction}{Which reduction should be the input.}

\item{dims.use}{How many dimensions of the reduction should be used?}

\item{nl.reduction.method}{Which non-linear reduction method should be applied? 'umap', 'tsne'}

\item{benchmark}{TRUE/FALSE whether benchmarking should be performed.}

\item{dist.method}{Which distance method should be used to calculate distance. Available: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'}
}
\description{
Includes: normalisation, scale/centre, unwanted variance regression, linear dimensionality reduction (PCA), and benchmarking.
This function iterates through all normalisation techniques and benchmarks to indicate the superior method.
}
