% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_data.R
\name{Scale_data}
\alias{Scale_data}
\title{Here scales, centres, and regresses unwanted confounders}
\usage{
Scale_data(object, unwanted.variance, scale, centre, ...)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{unwanted.variance}{Which column metadata you wish to regress, can be listed as c('factor_1', 'factore_2')}

\item{scale}{TRUE/FALSE whether to scale features}

\item{centre}{TRUE/FALSE whether to centre transform the matrix to make the mean equal 0}

\item{...}{Specify parameters to be passed to the ScaleData function.}
}
\description{
Here scales, centres, and regresses unwanted confounders
}
\examples{
Scale_data(object = sce_object, unwanted.variance = 'percent.mt', scale = TRUE, centre = TRUE)

}
