% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNormalise.R
\name{scNormalise}
\alias{scNormalise}
\title{An integrated normalisation step}
\usage{
scNormalise(
  object,
  method = "scran",
  scale,
  centre,
  unwanted.variance,
  n.features = 2000,
  benchmark,
  distance.method,
  scran.max.cluster.size = 1000,
  scran.scaling = NULL,
  scran.log = TRUE,
  scran.center_size_factors = TRUE
)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{method}{Specify 'all' to perform all methods and store them as alternative experiemnts}

\item{scale}{Should the data be scaled?}

\item{centre}{Should the data be centre transformed?}

\item{unwanted.variance}{Specify which metadata columns should be regressed.}

\item{n.features}{How many highly variable genes should be retained.}

\item{benchmark}{TRUE/FALSE whether benchmarking should be performed.}

\item{distance.method}{SpecWhich distance method should be used to calculate distance. Available: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'}

\item{...}{If only a single method is applied, you may pass parameters down to the selected function here.}
}
\description{
Includes: normalisation, scale/centre, unwanted variance regression, linear dimensionality reduction (PCA), and benchmarking.
This function iterates through all normalisation techniques and benchmarks to indicate the superior method.
}
