% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQC_correlate.R
\name{scQC_correlate}
\alias{scQC_correlate}
\title{Produces a 3D scatterplot displaying specific metadata information.}
\usage{
scQC_correlate(object, x, y, z)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{x}{x-axis}

\item{y}{y-axis}

\item{z}{colour gradient}
}
\description{
This function produces a plot with ggplots2, where three metadata variables can be visualised simultaneously.
The y-axis and x-axis will display two metadata variables, and the thrid dimension is introduced with a colour gradient.
}
\details{
Remember, you can view your available  metadata by specifying colData(sce_object)
}
