% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQC_density.R
\name{scQC_density}
\alias{scQC_density}
\title{Produces a density plot displaying specific metadata information.}
\usage{
scQC_density(object, column, title, cutoff)
}
\arguments{
\item{object}{Please specify a SCE object produced using IBRAP functions.}

\item{column}{Which metadata column would you like to}

\item{title}{What should the title be?}

\item{cutoff}{Where should a cutoff point be visualised. This is important for comprehending the effect of the filtration stage.}
}
\description{
This function produces a plot with ggplots2, where y correlates with n of cells and x with a metadata variable.
}
\details{
Remember, you can view your available  metadata by specifying colData(sce_object)
}
