% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showObjectContents.R
\name{showObjectContents}
\alias{showObjectContents}
\title{Shows the contents in your IBRAP object}
\usage{
showObjectContents(object, assay)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing which assays to show}
}
\value{
Prints out the contents of the supplied assays
}
\description{
Shows the contents in your IBRAP object
}
\examples{

}
