% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitIBRAP.R
\name{splitIBRAP}
\alias{splitIBRAP}
\title{Split the IBRAP object up}
\usage{
splitIBRAP(object, split.by)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{split.by}{Character. Which column contained within object.sample_metadata should be used to split the object up}
}
\value{
A list of the separated IBRAP objects
}
\description{
This breaks the IBRAP object up by a specified IBRAP object.
}
\examples{
list.of.objects <- splitIBRAP(object = object, 
                                        split.by = 'original.project')

}
