% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singelbracketsubset.R
\name{[,IBRAP,ANY,ANY,ANY-method}
\alias{[,IBRAP,ANY,ANY,ANY-method}
\title{Method override for `'['` subset function regarding IBRAP S4 object}
\usage{
\S4method{[}{IBRAP,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\description{
Method override for `'['` subset function regarding IBRAP S4 object
}
