% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracketsubset_IBRAP_method.R
\name{[[,IBRAP,ANY,ANY-method}
\alias{[[,IBRAP,ANY,ANY-method}
\title{Method override for subset method `'[['` for IBRAP object}
\usage{
\S4method{[[}{IBRAP,ANY,ANY}(x, i, j, ...)
}
\description{
Method override for subset method `'[['` for IBRAP object
}
