% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.clustering.R
\name{benchmark.clustering}
\alias{benchmark.clustering}
\title{Benchmarks the cluster assignmentws}
\usage{
benchmark.clustering(
  object,
  assay,
  clustering,
  reduction,
  n.dims = 1:3,
  dist.method = "euclidean",
  ground.truth = NULL
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{clustering}{Character. The names of the cluster assignment dataframes to use}

\item{reduction}{Character. Which reduction(s) within the assay should be supplied for distance calcultions}

\item{n.dims}{Numerical. How many dimensions of the reduction should be supplied. Default = 1:3}

\item{dist.method}{Character. Which distance method should be used, options: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary', 'minkowski'. Default = 'euclidean'}

\item{ground.truth}{Vector. If available, supply the vector in the same order as colnames of the ground truth, i.e. true cell type labels. If this is not supplied, only unsupervised methods will be supplied. Default = NULL}
}
\value{
Benchmarking scores for the supplied cluster assignments
}
\description{
Supervised (ARI and NMI) and unsupervised (ASW, Dunn Index, and Connectivity) benchmarking metrics are calculated for cluster assignments. assays, clustering and reduction for distance calculations are iterated through.
}
