% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_percentage_genes.R
\name{find_percentage_genes}
\alias{find_percentage_genes}
\title{Finds the most variable genes in a dataset}
\usage{
find_percentage_genes(object = obj, pattern = '^MT-', assay = 'RAW', slot = 'RAW', column.name = 'RAW_percent.mt')
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{pattern}{A character string containing a pattern to identfiy in rownames}

\item{assay}{String indicating which assay to source the raw counts from}

\item{slot}{String indicating which slot within the assay should be sourced}

\item{column.name}{String naming the column name in the metadata}
}
\value{
IBRAP S4 class object containing highly varaible genes within the source assay
}
\description{
Find an x amount of top variable genes for your dataset
}
