% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.harmony.R
\name{perform.harmony}
\alias{perform.harmony}
\title{Performs Harmony integration}
\usage{
perform.harmony(
  object,
  assay,
  vars.use,
  reduction = "pca",
  reduction.save = "harmony",
  dims.use = NULL,
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{vars.use}{Character. A string of the column nmae that contains variables to regress.}

\item{reduction}{Character. String defining the name of the reduction to provide for BBKNN. Default = 'pca'}

\item{reduction.save}{Character. What should the harmony reduction be saved as. Default = 'harmony'}

\item{dims.use}{Numerical. Number of dimensions of the provided reduction to input into harmony, NULL equates to all dimensions. Default = NULL}

\item{...}{Arguments to be passed to PCAtools::pca}
}
\value{
PCA reductions contained within the computational_reduction list in the defined assays
}
\description{
Performs Harmony integration on defined method-assays and reductions contained within. This is performed on reductions.
}
