% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.umap.R
\name{perform.umap}
\alias{perform.umap}
\title{Performs UMAP reduction}
\usage{
perform.umap(
  object,
  assay,
  reduction = NULL,
  graph = NULL,
  reduction.save = "umap",
  n.dims = NULL,
  n_components = 3,
  n_neighbors = 30,
  metric = "cosine",
  min_dist = 0.3,
  ...
)
}
\arguments{
\item{object}{IBRAP S4 class object}

\item{assay}{Character. String containing indicating which assay to use}

\item{reduction}{Character. String defining which reduction to supply to the UMAP algorithm. Default = NULL}

\item{graph}{Character. If you wish to UMAP project a previously created connectivity graph (i.e. BBKNN output) supply the graph name here with reductions set to NULL. Default = NULL}

\item{n.dims}{Numerical. The number of Scanorama dimensions to be produced. Default = 50}

\item{n_components}{Numerical. How many UMAP dimensions should be produced, if you are supplying graphs, only 2 dimensions can be produced. Default = 3}

\item{...}{Numerical. Arguments to be passed to Seurat::RunUMAP}
}
\value{
UMAP reduction saved in the visualisation_reductions section in the supplied method-assays
}
\description{
Performs UMAP reduction on defined method-assays and supplied reductions or graphs.
}
