% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GO.output.R
\name{plot.GO.output}
\alias{plot.GO.output}
\title{Plot GO enrichment output}
\usage{
\method{plot}{GO.output}(result)
}
\arguments{
\item{result}{The result file from the perform.GO.enrichment}
}
\description{
Creates a dot plot of GO terms
}
