% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_10X_output.R
\name{Read10X_output}
\alias{Read10X_output}
\title{Read 10x Cell Ranger output files}
\usage{
Read10X_output(directory)
}
\arguments{
\item{directory}{Paste the ABSOLUTE pathway to the file containing the three output files? Files must be named matrix.mtx, genes.tsv, and barcodes.tsv}
}
\description{
This function allows you concatenate the three matrix files into an expression matrix.
}
\examples{
Read10X_output('place directory here')

}
\keyword{10x}
\keyword{CellRanger}
